@@ -40,10 +40,9 @@
     "prepublish": "yarn run build"
   },
   "dependencies": {
-    "babel-runtime": "^6.26.0",
     "@babel/runtime-corejs2": "^7.7.7",
     "froala-editor": "^3.2.7",
-    "vue": "2.6.6"
+    "vue": "^3.0.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.2.3",
@@ -53,7 +52,7 @@
     "@babel/polyfill": "^7.2.5",
     "@babel/preset-env": "^7.3.4",
     "@babel/runtime": "^7.3.4",
-    "@babel/runtime-corejs2": "^7.3.4",
+    "@vue/compiler-sfc": "^3.0.7",
     "babel-helpers": "^6.24.1",
     "babel-loader": "^8.0.5",
     "connect-history-api-fallback": "^1.1.0",
@@ -70,7 +69,6 @@
     "inject-loader": "^4.0.1",
     "isparta-loader": "^2.0.0",
     "jasmine-core": "^3.3.0",
-    "json-loader": "^0.5.4",
     "lolex": "^3.1.0",
     "nightwatch": "^1.0.19",
     "ora": "^3.1.0",
@@ -79,9 +77,8 @@
     "url-loader": "^1.1.2",
     "vue-hot-reload-api": "^2.0.11",
     "vue-html-loader": "^1.2.4",
-    "vue-loader": "15.6.2",
+    "vue-loader": "^16.8.3",
     "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "2.6.6",
     "webpack": "^4.29.3",
     "webpack-cli": "3.2.3",
     "webpack-dev-middleware": "^3.5.2",
