@@ -5,12 +5,10 @@ import 'froala-editor/css/froala_editor.pkgd.min.css';
 import 'froala-editor/css/froala_style.min.css';
 import 'froala-editor/js/plugins.pkgd.min.js';
 
-import App from './examples/App'
-import Vue from 'vue'
+import App from './examples/App';
+import { createApp, h } from 'vue';
 import VueFroala from 'src';
 
-Vue.use(VueFroala);
-new Vue({
- render: h => h(App)
-
-}).$mount('#app')
+createApp({
+ render: () => h(App)
+}).use(VueFroala).mount('#app');
